+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+* correctly order replay frames and remove invalid frames. This changes similarity values slightly
+
+# v4.0.2
+
+* loading replays from cache is now roughly twice as fast
+
 # v4.0.1
 
 * fix `ReplayMap`s inside a `Map` not having some attributes set after the `Map` is info loaded